+++ .gitignore
@@ -79,6 +79,3 @@ lint/generated/
 lint/outputs/
 lint/tmp/
 # lint/reports/
-
-.DS_Store
-/.idea
