+++ FtcRobotController/src/main/AndroidManifest.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
           xmlns:tools="http://schemas.android.com/tools"
-    android:versionCode="52"
-          android:versionName="9.0.1">
+    android:versionCode="51"
+          android:versionName="9.0">
 
   <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
