+++ MeepMeepTesting/src/main/java/com/example/meepmeeptesting/MeepMeepTesting.java
@@ -1,0 +1,45 @@
+package com.example.meepmeeptesting;
+
+import com.acmerobotics.roadrunner.Pose2d;
+import com.acmerobotics.roadrunner.Vector2d;
+import com.noahbres.meepmeep.MeepMeep;
+import com.noahbres.meepmeep.roadrunner.DefaultBotBuilder;
+import com.noahbres.meepmeep.roadrunner.entity.RoadRunnerBotEntity;
+
+import java.awt.Image;
+import java.io.File;
+import java.io.IOException;
+
+import javax.imageio.ImageIO;
+
+public class MeepMeepTesting {
+    public static void main(String[] args) {
+        MeepMeep meepMeep = new MeepMeep(800);
+
+        RoadRunnerBotEntity myBot = new DefaultBotBuilder(meepMeep)
+                // Set bot constraints: maxVel, maxAccel, maxAngVel, maxAngAccel, track width
+                .setConstraints(50, 50, Math.toRadians(180), Math.toRadians(180), 10.276)
+                .build();
+
+        myBot.runAction(myBot.getDrive().actionBuilder(new Pose2d(0, 0, 0))
+                .splineTo(new Vector2d(30, 30), Math.PI / 2)
+                .waitSeconds(3)
+                .splineToLinearHeading(new Pose2d(60, 0,Math.PI), -Math.PI/2)
+                        .waitSeconds(1)
+                        .setTangent(Math.PI)
+                .splineToLinearHeading(new Pose2d(20, 0, 0), Math.PI)
+                        .lineToX(0)
+                .build());
+
+
+        Image img = null;
+        try { img = ImageIO.read(new File("./MeepMeepTesting/field-2023-official.png")); }
+        catch (IOException e) {}
+
+        meepMeep.setBackground(img)
+                .setDarkMode(true)
+                .setBackgroundAlpha(0.95f)
+                .addEntity(myBot)
+                .start();
+    }
+}
