+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/PoseMessage.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.hardware;
 
 import com.acmerobotics.roadrunner.Pose2d;
