+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/TwoDeadWheelLocalizer.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.hardware;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.DualNum;
@@ -8,7 +8,6 @@ import com.acmerobotics.roadrunner.Twist2dDual;
 import com.acmerobotics.roadrunner.Vector2dDual;
 import com.acmerobotics.roadrunner.ftc.Encoder;
 import com.acmerobotics.roadrunner.ftc.FlightRecorder;
-import com.acmerobotics.roadrunner.ftc.OverflowEncoder;
 import com.acmerobotics.roadrunner.ftc.PositionVelocityPair;
 import com.acmerobotics.roadrunner.ftc.RawEncoder;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
@@ -20,8 +19,8 @@ import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 @Config
 public final class TwoDeadWheelLocalizer implements Localizer {
     public static class Params {
-        public double parYTicks = 0.0; // y position of the parallel encoder (in tick units)
-        public double perpXTicks = 0.0; // x position of the perpendicular encoder (in tick units)
+        public double PAR_Y_TICKS = 0.0;
+        public double PERP_X_TICKS = 0.0;
     }
 
     public static Params PARAMS = new Params();
@@ -34,11 +33,9 @@ public final class TwoDeadWheelLocalizer implements Localizer {
 
     private final double inPerTick;
 
-    private double lastRawHeadingVel, headingVelOffset;
-
     public TwoDeadWheelLocalizer(HardwareMap hardwareMap, IMU imu, double inPerTick) {
-        par = new OverflowEncoder(new RawEncoder(hardwareMap.get(DcMotorEx.class, "par")));
-        perp = new OverflowEncoder(new RawEncoder(hardwareMap.get(DcMotorEx.class, "perp")));
+        par = new RawEncoder(hardwareMap.get(DcMotorEx.class, "par"));
+        perp = new RawEncoder(hardwareMap.get(DcMotorEx.class, "perp"));
         this.imu = imu;
 
         lastParPos = par.getPositionAndVelocity().position;
@@ -50,16 +47,6 @@ public final class TwoDeadWheelLocalizer implements Localizer {
         FlightRecorder.write("TWO_DEAD_WHEEL_PARAMS", PARAMS);
     }
 
-    // see https://github.com/FIRST-Tech-Challenge/FtcRobotController/issues/617
-    private double getHeadingVelocity() {
-        double rawHeadingVel = imu.getRobotAngularVelocity(AngleUnit.RADIANS).zRotationRate;
-        if (Math.abs(rawHeadingVel - lastRawHeadingVel) > Math.PI) {
-            headingVelOffset -= Math.signum(rawHeadingVel) * 2 * Math.PI;
-        }
-        lastRawHeadingVel = rawHeadingVel;
-        return headingVelOffset + rawHeadingVel;
-    }
-
     public Twist2dDual<Time> update() {
         PositionVelocityPair parPosVel = par.getPositionAndVelocity();
         PositionVelocityPair perpPosVel = perp.getPositionAndVelocity();
@@ -69,17 +56,17 @@ public final class TwoDeadWheelLocalizer implements Localizer {
         int perpPosDelta = perpPosVel.position - lastPerpPos;
         double headingDelta = heading.minus(lastHeading);
 
-        double headingVel = getHeadingVelocity();
+        double headingVel = imu.getRobotAngularVelocity(AngleUnit.RADIANS).zRotationRate;
 
         Twist2dDual<Time> twist = new Twist2dDual<>(
                 new Vector2dDual<>(
                         new DualNum<Time>(new double[] {
-                                parPosDelta - PARAMS.parYTicks * headingDelta,
-                                parPosVel.velocity - PARAMS.parYTicks * headingVel,
+                                parPosDelta - PARAMS.PAR_Y_TICKS * headingDelta,
+                                parPosVel.velocity - PARAMS.PAR_Y_TICKS * headingVel,
                         }).times(inPerTick),
                         new DualNum<Time>(new double[] {
-                                perpPosDelta - PARAMS.perpXTicks * headingDelta,
-                                perpPosVel.velocity - PARAMS.perpXTicks * headingVel,
+                                perpPosDelta - PARAMS.PERP_X_TICKS * headingDelta,
+                                perpPosVel.velocity - PARAMS.PERP_X_TICKS * headingVel,
                         }).times(inPerTick)
                 ),
                 new DualNum<>(new double[] {
