+++ build.dependencies.gradle
@@ -4,15 +4,15 @@ repositories {
 }
 
 dependencies {
-    implementation 'org.firstinspires.ftc:Inspection:9.0.1'
-    implementation 'org.firstinspires.ftc:Blocks:9.0.1'
-    implementation 'org.firstinspires.ftc:Tfod:9.0.1'
-    implementation 'org.firstinspires.ftc:RobotCore:9.0.1'
-    implementation 'org.firstinspires.ftc:RobotServer:9.0.1'
-    implementation 'org.firstinspires.ftc:OnBotJava:9.0.1'
-    implementation 'org.firstinspires.ftc:Hardware:9.0.1'
-    implementation 'org.firstinspires.ftc:FtcCommon:9.0.1'
-    implementation 'org.firstinspires.ftc:Vision:9.0.1'
+    implementation 'org.firstinspires.ftc:Inspection:9.0.0'
+    implementation 'org.firstinspires.ftc:Blocks:9.0.0'
+    implementation 'org.firstinspires.ftc:Tfod:9.0.0'
+    implementation 'org.firstinspires.ftc:RobotCore:9.0.0'
+    implementation 'org.firstinspires.ftc:RobotServer:9.0.0'
+    implementation 'org.firstinspires.ftc:OnBotJava:9.0.0'
+    implementation 'org.firstinspires.ftc:Hardware:9.0.0'
+    implementation 'org.firstinspires.ftc:FtcCommon:9.0.0'
+    implementation 'org.firstinspires.ftc:Vision:9.0.0'
     implementation 'org.firstinspires.ftc:gameAssets-CenterStage:1.0.0'
     implementation 'org.tensorflow:tensorflow-lite-task-vision:0.4.3'
     runtimeOnly 'org.tensorflow:tensorflow-lite:2.12.0'
