+++ settings.gradle
@@ -1,2 +1,4 @@
 include ':FtcRobotController'
 include ':TeamCode'
+include ':MeepMeepTesting'
+include ':MeepMeepTesting'
